#13 [ADD] UNKNOWN 타입 추가
