+++ core/src/main/java/com/gdg/core/type/DustConditionType.kt
@@ -3,6 +3,7 @@ package com.gdg.core.type
 import androidx.annotation.StringRes
 import androidx.compose.ui.graphics.Color
 import com.gdg.core.R
+import com.gdg.core.designsystem.theme.Gray700
 import com.gdg.core.designsystem.theme.Green600
 import com.gdg.core.designsystem.theme.Orange
 import com.gdg.core.designsystem.theme.Red
@@ -22,5 +23,9 @@ enum class DustConditionType(
     BAD(
         title = R.string.dust_condition_bad,
         color = Red
+    ),
+    UNKNOWN(
+        title = R.string.dust_condition_unknown,
+        color = Gray700
     )
 }
