+++ core/src/main/res/drawable/ic_congestion_gray.xml
@@ -1,0 +1,20 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="90dp"
+    android:height="90dp"
+    android:viewportWidth="90"
+    android:viewportHeight="90">
+  <path
+      android:pathData="M45,90C69.853,90 90,69.853 90,45C90,20.147 69.853,0 45,0C20.147,0 0,20.147 0,45C0,69.853 20.147,90 45,90Z"
+      android:strokeAlpha="0.594703"
+      android:fillColor="#808080"
+      android:fillType="evenOdd"
+      android:fillAlpha="0.594703"/>
+  <path
+      android:pathData="M45,61.694C54.22,61.694 61.694,54.22 61.694,45C61.694,35.78 54.22,28.306 45,28.306C35.781,28.306 28.307,35.78 28.307,45C28.307,54.22 35.781,61.694 45,61.694Z"
+      android:fillColor="#808080"
+      android:fillType="evenOdd"/>
+  <path
+      android:pathData="M45,50.806C48.207,50.806 50.806,48.207 50.806,45C50.806,41.793 48.207,39.194 45,39.194C41.793,39.194 39.193,41.793 39.193,45C39.193,48.207 41.793,50.806 45,50.806Z"
+      android:fillColor="#ffffff"
+      android:fillType="evenOdd"/>
+</vector>
