+++ feature/src/main/java/com/gdg/feature/detail/DetailViewModel.kt
@@ -10,6 +10,8 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
 import javax.inject.Inject
 
 @HiltViewModel
@@ -19,6 +21,10 @@ class DetailViewModel @Inject constructor(
     private val _getWeatherState: MutableStateFlow<UiState<WeatherEntity>> =
         MutableStateFlow(UiState.Empty)
     val getWeatherState: StateFlow<UiState<WeatherEntity>> get() = _getWeatherState
+
+    private val _getCongestionState: MutableStateFlow<UiState<CongestionEntity>> =
+        MutableStateFlow(UiState.Empty)
+    val getCongestionState: StateFlow<UiState<CongestionEntity>> get() = _getCongestionState
 
     fun getWeather(areaId: Long) {
         viewModelScope.launch {
@@ -34,22 +40,24 @@ class DetailViewModel @Inject constructor(
         }
     }
 
-    val mockWeather = WeatherEntity(
-        id = 1,
-        areaNm = "강남역",
-        skyStts = "구름많음",
-        temp = -3,
-        pm25Index = "좋음",
-        pm10Index = "나쁨",
-    )
+    fun getCongestion(areaId: Long) {
+        viewModelScope.launch {
+            _getCongestionState.emit(UiState.Loading)
+            crowdZeroRepository.getCongestion(areaId).fold(
+                onSuccess = {
+                    _getCongestionState.emit(UiState.Success(it))
+                },
+                onFailure = {
+                    _getCongestionState.emit(UiState.Failure(it.message.toString()))
+                }
+            )
+        }
+    }
 
-    val mockCongestion = CongestionEntity(
-        id = 1,
-        name = "강남역",
-        level = "보통",
-        message = "사람이 몰려있을 가능성이 낮고 붐빔은 거의 느껴지지 않아요\n도보 이동이 자유로워요",
-        min = 3800,
-        max = 4000,
-        time = "2025-02-01T10:00:00"
-    )
+    fun getCurrentTimeISO8601(): String {
+        val current = LocalDateTime.now()
+        val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss")
+        return current.format(formatter)
+    }
+
 }
