+++ feature/src/main/res/values/strings.xml
@@ -43,5 +43,6 @@
     <string name="calendar_people_reporting">%1$s명</string>
     <string name="calender_jurisdiction">관할시</string>
     <string name="calender_no_info">해당 날짜의 집회 정보가 없습니다</string>
+    <string name="detail_congestion_failure">실시간 인구 정보를 가져오지 못했어요</string>
 
 </resources>
